feat: API利用状況を利用量ページに移動、Discord追加、トップバー整理
- プロフィールからAPI利用状況を削除し、支払い > 利用量に移動
- ドキュメントとDiscordに外部リンクアイコンを追加
- サポートセクションにDiscordリンクを追加
- トップバーからドキュメントとDiscordアイコンを削除
